fill bug
This change fixes a bug filling intel hex files. It turns out that a number of filters, not just the fill filter, assume that the termination record is the last record in the file. For some formats this is not true, so they have been re-coded without this assumption. The "termination" record has been renamed the "start address" record, to more accurately refect what it does, and to remove the expectation that it is the last record in the file.

My thanks to Neil Viberg <Neil.Viberg@cmcelectronics.ca> for reporting this problem.