seek less on binary files
This change causes srec_cat to seek less when it outputs to binary files.  This was not a problem for regular files, but it made a mess when piping the standard output to some other program (you can't seek on pipes).  Most of the the the seeks were unnecessary, so they are avoided whenever possible.

My thanks to Stuart Warren <Stuart.WARREN@dds.delphiauto.com> for this suggestion.