option to emit CRLF
There is a new srec_cat --crlf option, which may be used for force CRLF output on operating systems which don't use that line termination.

My thanks to Adam Gibson for this suggestion:
"Some software that uploads firmware expect the hexfiles to be in CR/LF format.  To keep from needing another tool dependancy it would be nice if there were an option to output in CR/LF format."