Fix endless loop in generator bug
Marksu Heidelberg <markus.heidelberg@web.de> contributed a patch to fix a bug that he also found with the -generator which resulted in an endless loop.

If the start address of a partial record is >= 0xFFFFFFF1, there is a 32 overflow when adding 255 (the maximum record data length). But the result with this address is still correct (see documentation "If the maximum is given as zero then the range extends to the end of the address space."), the endless loop appears when using address 0xFFFFFFF2 or higher.

But can be previously reproduced with this command: srec_cat -generate 0xFFFFFF02 0xFFFFFF03 -constant 0xAB