srec_info: Fix verbose ration DivByZero
If the first and last byte of an 32-bit address range is used,
the size will be calculated as zero.
That size was used as divisor which needs to be avoided.

Additionally the number of bytes could be zero if all bytes are filled.
This caused the allocation result to be 0% instead of 100%.

Fixes #10