CI: Small improvements on GitHub workflow
Use `apt-get` instead of `apt`.
Avoids a warning that `apt` is not suitable for automation.

Use single installation call to avoid ignored exit code.

Add separate steps for smoke test and test preparation.

Remove permissions settings which were needed for the remove test report.

Explicit setting of shell is not necessary anymore, the tests run fine
on Ubuntu without it.
Maybe the addition of the executable flags or fixing the `shellcheck`
issues helped. These were done in previous commits.