[feature] Changed documentation generation behavior
Changed cmake variable for enabling documentation to `SRecord_ENABLE_DOCS` and
set default behavior to `OFF`.

Also modified how document generation is determined. If `SRecord_ENABLE_DOCS` is
true, then documentation will always be generated, otherwise documentation
will only be generated for a release build.

If documentation is to be generated and doxygen not found, then a fatal error
will be generated.